Minor improvements
Minor cleanup, add zobrist tests (castling_rights_affect_zobrist() does not pass yet)

+++ protochess-engine-rs/src/move_generator/attack_tables/mod.rs
@@ -150,7 +150,7 @@ impl AttackTables {
             if occ.get_bit(index as BIndex) {
                 occ_index |= 1;
             }
-            index += 15 as i16;
+            index += 15;
         }
         if y + x > 15 {
             occ_index <<= 16 - num_iters;
